fix(blade): cleanup
