+++ packages/blade/jest-setup.native.js
@@ -1,0 +1,3 @@
+Object.defineProperty(global.navigator, 'product', {
+  value: 'ReactNative',
+});
