+++ packages/blade/src/utils/mocks/setupMatchMediaMock.ts
@@ -1,0 +1,14 @@
+const setupMatchMediaMock = (customArgs = {}): void => {
+  Object.defineProperty(window, 'matchMedia', {
+    writable: true,
+    value: jest.fn().mockImplementation((query) => ({
+      matches: false,
+      media: query,
+      addEventListener: jest.fn(),
+      removeEventListener: jest.fn(),
+      ...customArgs,
+    })),
+  });
+};
+
+export default setupMatchMediaMock;
